+++ src/main/java/edu/fiuba/algo3/controllers/ControladoresTienda/ControladorBotonAgregar.java
@@ -2,10 +2,10 @@ package edu.fiuba.algo3.controllers.ControladoresTienda;
 
 import edu.fiuba.algo3.modelo.Balatro;
 import edu.fiuba.algo3.modelo.Carta.Carta;
-import edu.fiuba.algo3.modelo.EspacioJokersLlenoExcepction;
+import edu.fiuba.algo3.modelo.Excepcion.EspacioJokersLlenoExcepction;
 import edu.fiuba.algo3.modelo.Joker.Joker;
 import edu.fiuba.algo3.vistas.AvisosYAlertas.AlertaLimiteJoker;
-import edu.fiuba.algo3.vistas.VistaPartida;
+import edu.fiuba.algo3.vistas.VistasAuxiliares.VistaPartida;
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 import javafx.scene.layout.StackPane;
