+++ src/main/java/edu/fiuba/algo3/modelo/Excepcion/EspacioJokersLlenoExcepction.java
@@ -1,4 +1,4 @@
-package edu.fiuba.algo3.modelo;
+package edu.fiuba.algo3.modelo.Excepcion;
 
 public class EspacioJokersLlenoExcepction extends RuntimeException {
     public EspacioJokersLlenoExcepction(String message) {
