+++ src/main/java/edu/fiuba/algo3/vistas/App.java
@@ -1,31 +1,17 @@
 package edu.fiuba.algo3.vistas;
 
 import edu.fiuba.algo3.modelo.Balatro;
-import edu.fiuba.algo3.modelo.Carta.Carta;
-import edu.fiuba.algo3.modelo.Mazo;
-import edu.fiuba.algo3.modelo.Mezclador.IMezclador;
-import edu.fiuba.algo3.modelo.Mezclador.MezcladorMazo;
 import edu.fiuba.algo3.modelo.parser.ParserBalatro;
+import edu.fiuba.algo3.vistas.ElementosAuxiliares.ReproductorMusica;
+import edu.fiuba.algo3.vistas.VistasAuxiliares.VistaMenuInicial;
 import javafx.application.Application;
-import javafx.application.Platform;
-import javafx.geometry.Insets;
-import javafx.geometry.Pos;
 import javafx.scene.Scene;
-import javafx.scene.control.*;
 import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-import javafx.scene.layout.*;
 import javafx.scene.text.Font;
 import javafx.stage.Stage;
 
-import javafx.scene.media.Media;
-import javafx.scene.media.MediaPlayer;
-
 
 import java.io.FileNotFoundException;
-import java.util.ArrayList;
-import java.util.Objects;
-import java.util.Optional;
 
 /**
  * JavaFX App
