+++ src/main/java/edu/fiuba/algo3/vistas/ElementosAuxiliares/ReproductorMusica.java
@@ -1,4 +1,4 @@
-package edu.fiuba.algo3.vistas;
+package edu.fiuba.algo3.vistas.ElementosAuxiliares;
 
 import javafx.scene.media.Media;
 import javafx.scene.media.MediaPlayer;
