+++ src/main/java/edu/fiuba/algo3/vistas/VistasAuxiliares/VistaPopupDerrota.java
@@ -1,26 +1,25 @@
-package edu.fiuba.algo3.vistas;
+package edu.fiuba.algo3.vistas.VistasAuxiliares;
 
 import edu.fiuba.algo3.controllers.ControladorCierreJuego;
 import edu.fiuba.algo3.controllers.ControladorFinJuego;
+import edu.fiuba.algo3.vistas.ElementosAuxiliares.CrearBoton;
+import edu.fiuba.algo3.vistas.ElementosAuxiliares.ReproductorMusica;
+import javafx.geometry.Insets;
+import javafx.geometry.Pos;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
 import javafx.scene.layout.StackPane;
 import javafx.scene.layout.VBox;
-import javafx.geometry.Insets;
-import javafx.geometry.Pos;
-import javafx.scene.media.MediaPlayer;
 import javafx.stage.Stage;
 
-
-public class VistaPopupVictoria {
-
+public class VistaPopupDerrota {
     private Stage stage;
-    private ReproductorMusica reproductorMusica;
+    private ReproductorMusica reproductor;
     private String nombreJugador;
 
-    public VistaPopupVictoria(Stage stage, ReproductorMusica reproductorMusica, String nombreJugador) {
+    public VistaPopupDerrota(Stage stage, ReproductorMusica reproductor, String nombreJugador) {
         this.stage = stage;
-        this.reproductorMusica = reproductorMusica;
+        this.reproductor = reproductor;
         this.nombreJugador = nombreJugador;
     }
 
@@ -37,7 +36,7 @@ public class VistaPopupVictoria {
         recuadroPopup.setMaxWidth(500);
         recuadroPopup.setMaxHeight(300);
 
-        Label labelMensaje = new Label("Ganaste el juego, ¡felicitaciones!");
+        Label labelMensaje = new Label("Fin del juego, no has alcanzado el puntaje necesario");
         labelMensaje.setStyle("-fx-font-size: 16px; -fx-font-weight: bold; -fx-text-alignment: center; -fx-text-fill: white;");
 
         String estiloBoton = "-fx-background-color: #2C3E50; " +
@@ -52,7 +51,7 @@ public class VistaPopupVictoria {
                 "-fx-background-radius: 5px;";
 
         Button botonContinuar = new CrearBoton("Volver al menu principal", estiloBoton);
-        botonContinuar.setOnAction(new ControladorFinJuego(stage, reproductorMusica, nombreJugador));
+        botonContinuar.setOnAction(new ControladorFinJuego(stage, reproductor, nombreJugador));
 
         Button botonCerrar = new CrearBoton("Cerrar juego", estiloBoton);
         botonCerrar.setOnAction(new ControladorCierreJuego());
@@ -65,3 +64,4 @@ public class VistaPopupVictoria {
         rootPrincipal.getChildren().add(overlay);
     }
 }
+
