+++ src/main/java/module-info.java
@@ -8,4 +8,7 @@ module edu.fiuba.algo3 {
     exports edu.fiuba.algo3.vistas.VistasTienda;
     exports edu.fiuba.algo3.vistas.AvisosYAlertas;
     exports edu.fiuba.algo3.vistas.VistaMesaJuego;
+    exports edu.fiuba.algo3.vistas.Cajas;
+    exports edu.fiuba.algo3.vistas.ElementosAuxiliares;
+    exports edu.fiuba.algo3.vistas.VistasAuxiliares;
 }
